test report visual session service on electron
