+++ electron/app/session-services/test-report-visual.sm/index.ts
@@ -1,0 +1,1 @@
+export { TestReportVisualServiceModel } from './test-report-visual.sm';
